Rename S0142/process_raw -> S0142/processed_S0142
