+++ tests/ma/elexon/metering_data/test_transforms.py
@@ -1,7 +1,7 @@
 import data.register
 from ma.elexon.metering_data.metering_data_by_half_hour_and_bmu import MeteringDataHalfHourlyByBmu
 from ma.elexon.metering_data.metering_data_by_time import MeteringDataDaily, MeteringDataHalfHourly, MeteringDataMonthly
-from ma.elexon.S0142.process_raw import ProcessedS0142
+from ma.elexon.S0142.processed_S0142 import ProcessedS0142
 
 
 def test_transforms() -> None:
